https://github.com/apache/incubator-seatunnel/issues/2893
feat #2893 :
(1), solve the problem that the -i parameter cannot be passed in the placeholder of the source part of the configuration file that replaces the conf, and supports placeholder replacement like the sql part.

(2), solve druid's start_date, end_date through the command line -i to pass the parameter space, recognized as the parameter separator of the shell command line, so that the date format parameter of yyyy-MM-dd HH:mm:ss cannot be recognized as a whole, resulting in running Errors, use specific placeholders by default, can be customized

(3), customize the output separator of druid to support importing into doris

(4), replace the default delimiter of doris to solve the problem that " "(\t) in the column value affects column recognition.

(5), druid adds user, password parameter authentication support

(6), provide provided annotation for the package that the druid connector can rely on seatunnel