[Bug] [seatunnel-formats] Allow the entry in the map to be null and allow the key in the entry to be null (#5277)
* :bug: "Add null check in text deserialization

This commit adds a null check while splitting the kv in the TextDeserializationSchema. Not having this check could potentially lead to a crash if 'kvs' has less than two elements. Now, the function will return null in this case, significantly improving reliability."

* :bug: Refactor text deserialization, allow null value assignments

Modified the TextDeserializationSchema to allow key-value pairs with null values. This adjustment helps to handle data records with missing values more accurately. The code was also reformatted for improved readability.

* :art: Refactor Builder constructor in TextDeserializationSchema

Tidied up the Builder constructor in TextDeserializationSchema by removing unnecessary empty lines. This change improves code readability and consistency with the rest of the class?

* :white_check_mark: Add  test case where the value is empty and the key is empty in the map

---------

Co-authored-by: yangpeng <nickyooungp@gmail.com>