:art: Refactor HiveJdbcConnectionProvider and SimpleJdbcConnectionProvider classes
Changed the method of accessing JdbcConfig object and the Connection object in both classes from getter methods to direct access. This was done to create a more cohesive codebase and promote a uniform way of accessing these objects. This change also aids in improving the code readability and reducing ambiguities when dealing with these objects across different packages.