Add Hive JDBC dependency and update SimpleJdbcConnectionProvider
This commit introduces the Apache Hive JDBC connector as a dependency in the POM.xml to support connectivity with Hive. It also removes @Data annotation from SimpleJdbcConnectionProvider and manually implements getters and setters. This was done to improve code clarity and ensure explicit control over the modifications to Connection object.