[BugFix]: fix bug of source cassandra cause nullpointer exception
issue: https://github.com/apache/seatunnel/issues/9849

we do not need to throw exception when source field's content is null,
we need to sync data to sink, and the data in sink is also be null.

when the data is not null, we can convert
when the data is null, no need to convert, just sink null.